+++ presenter-compose/impl/src/androidMain/kotlin/software/ralf/app/platform/presenter/compose/AndroidComposePresenterScopeFactory.kt
@@ -3,6 +3,7 @@ package software.ralf.app.platform.presenter.compose
 import app.cash.molecule.AndroidUiDispatcher
 import app.cash.molecule.RecompositionMode
 import dev.zacsweers.metro.AppScope as MetroAppScope
+import dev.zacsweers.metro.BindingContainer
 import dev.zacsweers.metro.ContributesTo as MetroContributesTo
 import dev.zacsweers.metro.Provides as MetroProvides
 import dev.zacsweers.metro.SingleIn as MetroSingleIn
@@ -39,8 +40,8 @@ public interface AndroidComposePresenterScopeFactoryComponent {
 
 /** Provides the [AndroidComposePresenterScopeFactory] in the Metro graph. */
 @MetroContributesTo(MetroAppScope::class)
-@Suppress("CONTRIBUTES_TO_COULD_BE_BINDING_CONTAINER")
-public interface AndroidComposePresenterScopeFactoryGraph {
+@BindingContainer
+public object AndroidComposePresenterScopeFactoryGraph {
   /** Provides the [AndroidComposePresenterScopeFactory] in the Metro graph as a singleton. */
   @MetroProvides
   @MetroSingleIn(MetroAppScope::class)