+++ presenter-compose/impl/src/commonMain/kotlin/software/ralf/app/platform/presenter/compose/backgesture/DefaultBackGestureDispatcherPresenter.kt
@@ -1,6 +1,7 @@
 package software.ralf.app.platform.presenter.compose.backgesture
 
 import dev.zacsweers.metro.AppScope as MetroAppScope
+import dev.zacsweers.metro.BindingContainer
 import dev.zacsweers.metro.ContributesTo as MetroContributesTo
 import dev.zacsweers.metro.Provides as MetroProvides
 import dev.zacsweers.metro.SingleIn as MetroSingleIn
@@ -29,8 +30,8 @@ public interface DefaultBackGestureDispatcherPresenterComponent {
  * [BackGestureDispatcherPresenter] for more details.
  */
 @MetroContributesTo(MetroAppScope::class)
-@Suppress("CONTRIBUTES_TO_COULD_BE_BINDING_CONTAINER")
-public interface DefaultBackGestureDispatcherPresenterGraph {
+@BindingContainer
+public object DefaultBackGestureDispatcherPresenterGraph {
   /** Provides a [BackGestureDispatcherPresenter] as singleton in the Metro graph. */
   @MetroProvides
   @MetroSingleIn(MetroAppScope::class)