+++ presenter-compose/impl/src/linuxMain/kotlin/software/ralf/app/platform/presenter/compose/LinuxComposePresenterScopeFactory.kt
@@ -2,6 +2,7 @@ package software.ralf.app.platform.presenter.compose
 
 import app.cash.molecule.RecompositionMode
 import dev.zacsweers.metro.AppScope as MetroAppScope
+import dev.zacsweers.metro.BindingContainer
 import dev.zacsweers.metro.ContributesTo as MetroContributesTo
 import dev.zacsweers.metro.Provides as MetroProvides
 import dev.zacsweers.metro.SingleIn as MetroSingleIn
@@ -35,8 +36,8 @@ public interface LinuxComposePresenterScopeFactoryComponent {
 
 /** Provides the [LinuxComposePresenterScopeFactory] in the Metro graph. */
 @MetroContributesTo(MetroAppScope::class)
-@Suppress("CONTRIBUTES_TO_COULD_BE_BINDING_CONTAINER")
-public interface LinuxComposePresenterScopeFactoryGraph {
+@BindingContainer
+public object LinuxComposePresenterScopeFactoryGraph {
   /** Provides the [LinuxComposePresenterScopeFactory] in the Metro graph as a singleton. */
   @MetroProvides
   @MetroSingleIn(MetroAppScope::class)